Adding optional overlay when toggled to Rearview
When toggled show Rearview, the user may accidentally tap a control on
partially visible front view. This leads to undesired behaviour.

Fix:

Adding an optional overlay view. This view covers the front view
temporarily with a translucent black overlay, indicating that the view
is currently disabled. A button is added on top of this view which
toggles the views back.